requirements.txt
absl-py==0.7.1
aioredis==1.2.0
asgi-redis==1.4.3
asgiref==1.1.2
asn1crypto==0.24.0
astor==0.8.0
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.7.1
Automat==0.7.0
backcall==0.1.0
bleach==3.1.0
boto3==1.9.189
botocore==1.12.189
certifi==2019.3.9
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
cloudpickle==1.2.1
colorama==0.4.1
constantly==15.1.0
crispy-forms-semantic-ui==0.2.0
cryptography==2.7
cycler==0.10.0
daphne==2.3.0
decorator==4.4.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.2
django-ckeditor==5.7.1
django-crispy-forms==1.7.2
django-email-utils==0.1.1
django-filebrowser==3.11.1
django-grappelli==2.12.3
django-heroku==0.3.1
django-js-asset==1.2.2
django-markdownx==2.0.28
django-picklefield==1.0.0
django-rest-email-auth==2.0.2
django-semanticui-form==0.0.1
django-semanticui-forms==1.6.5
django-storages==1.7.1
django-templated-mail==1.1.1
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
djoser==2.0.3
docutils==0.14
entrypoints==0.3
future==0.17.1
gast==0.2.2
google-pasta==0.1.7
graphistry==0.9.64
grpcio==1.22.0
gunicorn==19.9.0
gym==0.13.1
h5py==2.9.0
hiredis==1.0.0
httpie==1.0.3
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.20
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.0.1
lazy-object-proxy==1.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
msgpack==0.6.1
msgpack-python==0.5.6
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.3
nltk==3.4.4
notebook==5.7.8
numpy==1.16.2
oauthlib==3.0.2
opencv-python==4.1.0.25
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pickleshare==0.7.5
Pillow==6.1.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
psycopg2==2.8.3
PyAudio==0.2.11
pycparser==2.19
pyglet==1.3.2
Pygments==2.3.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.0
pypiwin32==223
pyrsistent==0.14.11
PySocks==1.7.0
python-dateutil==2.8.0
python-decouple==3.1
python3-openid==3.1.0
pyttsx==1.1
pyttsx3==2.71
pytube==9.5.1
pytz==2019.1
pywin32==224
pywinpty==0.5.5
PyYAML==5.1.1
pyzmq==18.0.1
redis==2.10.6
regex==2019.6.8
requests==2.21.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
scikit-learn==0.21.2
scipy==1.2.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
snap==0.5
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
tweepy==3.8.0
Twisted==19.2.1
txaio==18.8.1
typed-ast==1.3.5
urllib3==1.24.1
vaderSentiment==3.2.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
whitenoise==4.1.3
widgetsnbextension==3.4.2
wrapt==1.11.1
zope.interface==4.6.0